+++ ext/Server/Bot.lua
@@ -413,11 +413,7 @@ function Bot:GetAttackDistance(p_ShootBackAfterHit, p_VehicleAttackMode)
 		end
 	else
 		if m_Vehicles:IsGunship(self.m_ActiveVehicle) then
-			if p_ShootBackAfterHit then
-				s_AttackDistance = Config.MaxShootDistanceGunship
-			else
-				s_AttackDistance = Config.MaxShootDistanceGunship
-			end
+			s_AttackDistance = Config.MaxShootDistanceGunship
 		elseif not m_Vehicles:IsAirVehicle(self.m_ActiveVehicle)
 			and m_Vehicles:IsNotVehicleType(self.m_ActiveVehicle, VehicleTypes.MobileArtillery)
 			and not m_Vehicles:IsAAVehicle(self.m_ActiveVehicle) then