+++ ext/Shared/Config.lua
@@ -97,6 +97,7 @@ Config = {
 	FovVerticleChopperForShooting = 80,	-- Degrees of pitch a chopper attacks 
 	FovVehicleAAForShooting = 360,		-- Degrees of FOV of AA - Vehicles 
 	FovVerticleVehicleAAForShooting = 160,	-- Degrees of FOV of AA - Vehicles 
+	MaxShootDistanceGunship = 700,		-- Meters bots in the Gunship start shooting at players 
 	MaxShootDistanceVehicles = 375,		-- Meters bots in Vehicles start shooting at players 
 	MaxShootDistanceGunship = 1000,		-- Meters bots in the Gunship starts shooting at players 
 	MaxShootDistanceNoAntiAir = 175,	-- Meters bots in vehicle (no Anti-Air) starts shooting at players